Use int64_t instead of int because of overflow. Solve day 19 part 2.

+++ 2024/day19/day19.h
@@ -5,9 +5,9 @@
 
 namespace day19 {
     int part1(const std::vector<std::string>& rows);
-    int part2(const std::vector<std::string>& rows);
+    int64_t part2(const std::vector<std::string>& rows);
 
-    int create_design(const std::string& design, const std::vector<bool>& towel_codes);
+    int64_t create_design(const std::string& design, const std::vector<bool>& towel_codes);
     std::vector<bool> get_towel_codes(const std::vector<std::string>& towels);
     int get_color_code(const char c);
 }
